Added the integrated code.

+++ download_data.py
@@ -6,8 +6,9 @@ import os
 url = "https://drive.google.com/uc?export=download&confirm=KdYt&id=16drOJ7K-B37bR2VpvxEvauUL4qgEgzjy"
 gdown.download(url, "data.tar.gz", quiet=False)
 
-if os.path.isdir("data"):
-    os.system('rm -r data')
-os.system("tar -xvzf data.tar.gz")
-os.system("mv output data")
-os.system("rm data.tar.gz")
+if os.name == 'posix':
+    if os.path.isdir("data"):
+        os.system('rm -r data')
+    os.system("tar -xvzf data.tar.gz")
+    os.system("mv output data")
+    os.system("rm data.tar.gz")
